BUG Preserve dtype when handling missing classes
np.full defaults to using `dtype=np.array(fill_value).dtype`, which for "predict_proba" (fill_value=0) is an int. Make the empty array here have the same data type as the predictions it's meant to hold.

+++ sklearn/model_selection/_validation.py
@@ -843,7 +843,8 @@ def _enforce_prediction_order(classes, predictions, n_classes, method):
                           'predict_proba': 0}
         predictions_for_all_classes = np.full((_num_samples(predictions),
                                                n_classes),
-                                              default_values[method])
+                                              default_values[method],
+                                              dtype=predictions.dtype)
         predictions_for_all_classes[:, classes] = predictions
         predictions = predictions_for_all_classes
     return predictions
